requirements.txt
fastapi~=0.115.12
requests~=2.32.3
beautifulsoup4~=4.13.4
readability-lxml~=0.8.4.1
tiktoken~=0.9.0
pydantic~=2.11.4
python-dotenv~=1.1.0
SQLAlchemy~=2.0.41
uvicorn[standard]
aiosqlite
greenlet